+++ src/main/kotlin/org/stellar/walletsdk/util/SponsorOperation.kt
@@ -1,4 +1,4 @@
-package org.stellar.walletsdk.utils
+package org.stellar.walletsdk.util
 
 import org.stellar.sdk.BeginSponsoringFutureReservesOperation
 import org.stellar.sdk.EndSponsoringFutureReservesOperation
