+++ src/main/kotlin/org/stellar/walletsdk/util/ValidateTransaction.kt
@@ -1,4 +1,4 @@
-package org.stellar.walletsdk.utils
+package org.stellar.walletsdk.util
 
 import org.stellar.sdk.Server
 import org.stellar.sdk.Transaction
