+++ src/test/kotlin/org/stellar/walletsdk/helpers/GetFileFromResource.kt
