+++ src/test/kotlin/org/stellar/walletsdk/util/AccountReservedBalanceTest.kt
@@ -1,4 +1,4 @@
-package org.stellar.walletsdk.utils
+package org.stellar.walletsdk.util
 
 import kotlin.test.assertEquals
 import org.junit.jupiter.api.DisplayName
@@ -6,21 +6,21 @@ import org.junit.jupiter.api.Test
 import org.stellar.sdk.responses.AccountResponse
 import org.stellar.walletsdk.helpers.objectFromJsonFile
 
-@DisplayName("accountAvailableNativeBalance")
-internal class AccountAvailableNativeBalanceTest {
+@DisplayName("accountReservedBalance")
+internal class AccountReservedBalanceTest {
   @Test
-  fun `basic funded account has 0 available balance`() {
+  fun `basic funded account`() {
     val account = objectFromJsonFile("account_basic.json", AccountResponse::class.java)
-    val availableBalance = accountAvailableNativeBalance(account)
+    val reservedBalance = accountReservedBalance(account)
 
-    assertEquals("0", availableBalance)
+    assertEquals(1.0, reservedBalance.toDouble())
   }
 
   @Test
   fun `more complex account`() {
     val account = objectFromJsonFile("account_full.json", AccountResponse::class.java)
-    val availableBalance = accountAvailableNativeBalance(account)
+    val reservedBalance = accountReservedBalance(account)
 
-    assertEquals("36.3598684", availableBalance)
+    assertEquals(6.5, reservedBalance.toDouble())
   }
 }
