+++ src/test/kotlin/org/stellar/walletsdk/util/ConvertLumensStroops.kt
@@ -1,4 +1,4 @@
-package org.stellar.walletsdk.utils
+package org.stellar.walletsdk.util
 
 import kotlin.test.assertEquals
 import org.junit.jupiter.api.DisplayName
