+++ src/test/kotlin/org/stellar/walletsdk/util/SponsorOperationTest.kt
@@ -1,4 +1,4 @@
-package org.stellar.walletsdk.utils
+package org.stellar.walletsdk.util
 
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
